do not copy build folder in webpack dockerfile
A new dev environment won't have the build folder in place until
webpack runs the build once. Without the build folder, the COPY
command in webpack's dockerfile fails - and so webpack can never
create the initial build. Since the build folder is only needed as
a volume mount, we remove it from the dockerfile.